+++ footer.php
@@ -63,6 +63,5 @@
 </div><!-- #page -->
 
 <?php wp_footer(); ?>
-	<script type="text/javascript" src="//cdn.jsdelivr.net/npm/slick-carousel@1.8.1/slick/slick.min.js"></script>
 </body>
 </html>