+++ views/content-search.php
@@ -10,31 +10,26 @@
 use GoodGuitarist\Custom\PostTypes;
 $the_post = get_post();
 $ypt_atts = PostTypes::get_block_attributes_from_post_content( $the_post->post_content, 'gutenberg-good-guitarist/ypt' );
+$post_thumbnail = get_template_directory_uri() . "/assets/dist/images/good-guitarist-preview-img.png";
+
+if ( has_post_thumbnail() ) {
+	$post_thumbnail = get_the_post_thumbnail();
+}
+if ( 'youtube-post' === get_post_type() ) {
+	$post_thumbnail = $ypt_atts['videoThumbnail'];
+}
 ?>
 
 <article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
-	<?php if ( 'youtube-post' === get_post_type() ): ?>
-		<a class="ypt-link-wrapper" aria-label="<?php the_title(); ?>" href="<?php echo esc_url( the_permalink() ); ?>">
-			<img src="<?php echo esc_url( $ypt_atts['videoThumbnail'] ); ?>">
-			<h2 class="entry-title"><?php the_title(); ?></h2>
+	<img src="<?php echo esc_url( $post_thumbnail ); ?>">
+	<div class="search-result-text">
+		<a class="link-wrapper"
+		aria-label="<?php the_title(); ?>"
+		href="<?php echo esc_url( the_permalink() ); ?>">
+		<?php the_title(); ?>
 		</a>
-	<?php else: ?>
-
-	<?php endif; ?>
-
-	<div class="entry-content">
-		<?php
-			the_excerpt();
-			wp_link_pages(
-				array(
-					'before' => '<div class="page-links">' . esc_html__( 'Pages:', 'good-guitarist' ),
-					'after'  => '</div>',
-				)
-			);
-		?>
-	</div><!-- .entry-content -->
-
-	<footer class="entry-footer">
-		<?php awps\core\tags::entry_footer(); ?>
-	</footer><!-- .entry-footer -->
-</article><!-- #post-## -->
+		<div class="entry-content">
+			<?php the_excerpt(); ?>
+		</div>
+	</div>
+</article>