+++ views/partials/song-and-artist.php
@@ -1,5 +1,5 @@
 <h3 class="song-and-artist">
-    <span><?php esc_html_e( $artist['name'] ?? '' ); ?></span>
+    <a href="<?php echo esc_url( $artist['url'] ); ?>"><?php esc_html_e( $artist['name'] ?? '' ); ?></a>
     <?php if ( ( $artist['name'] ?? false ) && ( $atts['songTitle'] ?? false ) ): ?>
         <span>-</span>
     <?php endif; ?>